chore: fixed some codestyle issues
