+++ src/web/src/components/playlist/PlaylistContext.tsx
@@ -3,4 +3,4 @@ import React from 'react';
 export const PlayListContext = React.createContext(null);
 
 
-//TODO: Fix so context keeps track of list and available droppable things
+// TODO: Fix so context keeps track of list and available droppable things
